update MNNVLinkTopologyInfo variable type
Situation:
- Previously, the data type is defined based on EM.
now we plan to use MNNVLinkTopologyInfo interface from PDI
and the type is defined to uint32_t

Change:
- update the data type to meet the PDI requirement

Fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-13222

+++ redfish-core/include/utils/nvidia_processor_utils.hpp
@@ -1617,10 +1617,10 @@ inline void getMNNVLinkTopologyInfo(
         const std::string* traySerialNumber = nullptr;
         const std::string* systemGUID = nullptr;
         const std::string* peerType = nullptr;
-        const uint64_t* moduleID = nullptr;
-        const uint64_t* hostID = nullptr;
-        const uint64_t* traySlotIndex = nullptr;
-        const uint64_t* traySlotNumber = nullptr;
+        const uint32_t* moduleID = nullptr;
+        const uint32_t* hostID = nullptr;
+        const uint32_t* traySlotIndex = nullptr;
+        const uint32_t* traySlotNumber = nullptr;
 
         const bool success = sdbusplus::unpackPropertiesNoThrow(
             dbus_utils::UnpackErrorPrinter(), resp, "ChassisSerialNumber",
